Cleanup to use Maven structure.
- removed eclipse-related configuration files.
- created a POM file to build the project with.
- removed binary libraries and added as maven dependencies
