+++ .gitignore
@@ -1,0 +1,5 @@
+bin
+.metadata
+sql-to-json.iml
+.idea/
+target/
