--- SqlToJson/.classpath
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
-	<classpathentry kind="lib" path="src/lib/mysql-connector-java-5.1.24-bin.jar"/>
-	<classpathentry kind="lib" path="src/lib/jackson-core-2.1.4.jar"/>
-	<classpathentry kind="lib" path="src/lib/jackson-databind-2.1.0.jar"/>
-	<classpathentry kind="lib" path="src/lib/jackson-annotations-2.1.4.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
