--- SqlToJson/config.json
@@ -1,37 +1,0 @@
-{
-    "jdbcDriver": "mysql",
-    "host": "localhost",
-    "database": "employees",
-    "user": "root",
-    "password": "",
-    "docRoot": "emp_no",
-    "query": "Select * FROM employees e left join titles t on e.emp_no = t.emp_no left join salaries s on e.emp_no = s.emp_no left join dept_emp de on de.emp_no = e.emp_no left join departments d on d.dept_no = de.dept_no left join dept_manager dm on dm.emp_no = e.emp_no",
-    "subNode": [
-        {
-            "startColumn": 8,
-            "endColumn": 10,
-            "name": "titles"
-        },
-        {
-            "startColumn": 12,
-            "endColumn": 14,
-            "name": "salaries"
-        },
-        {
-            "startColumn": 16,
-            "endColumn": 18,
-            "name": "dept_emp"
-        },
-        {
-            "startColumn": 19,
-            "endColumn": 20,
-            "name": "departments"
-        },
-        {
-            "startColumn": 21,
-            "endColumn": 24,
-            "name": "dept_manager"
-        }
-    ],
-    "prettyPrint": true
-}
