+++ src/test/resources/json/config.json.sample
@@ -1,0 +1,47 @@
+{
+	//Your DBMS types. Currently only support MySQL. Other coming soon
+    "jdbcDriver": "mysql",
+    "host": "localhost",
+    "database": "employees",
+    "user": "root",
+    "password": "",
+    "query": "Select * FROM employees e left join titles t on e.emp_no = t.emp_no left join salaries s on e.emp_no = s.emp_no left join dept_emp de on de.emp_no = e.emp_no left join departments d on d.dept_no = de.dept_no left join dept_manager dm on dm.emp_no = e.emp_no",
+    
+    //Document Root. A single identifier which will differentiate between JSON Object.
+    //It is the main table Primary Key that joined with other table's Foreign Key
+    "docRoot": "emp_no",
+    
+    //If you want to define certain columns from the ResultSet 
+    //as child / sub objects define it here
+    //The column start from 1. So the first is 1, the second is 2, etc...
+    "subNode": [
+        {
+            "startColumn": 8,
+            "endColumn": 10,
+            "name": "titles"
+        },
+        {
+            "startColumn": 12,
+            "endColumn": 14,
+            "name": "salaries"
+        },
+        {
+            "startColumn": 16,
+            "endColumn": 18,
+            "name": "dept_emp"
+        },
+        {
+            "startColumn": 19,
+            "endColumn": 20,
+            "name": "departments"
+        },
+        {
+            "startColumn": 21,
+            "endColumn": 24,
+            "name": "dept_manager"
+        }
+    ],
+    
+    //If you'd like to print the output easy to read, choose true
+    "prettyPrint": true
+}
